Commit Update Task Feature #906 Master dan Payroll :: FSD periode waktu
kerja (kalkulasi) :

- Modified checking employee status, if employee that still don't have
WtWorkingGroup AND status <> EMP_TERMINATION found, cancel the process.

- Add Checking before process, only allow if Log Finger Interface
process, and Finger Realization process already executed.

- Fixing Some Bug

+++ inkubator-hrm/src/main/java/com/inkubator/hrm/dao/impl/LogWtAttendanceRealizationDaoImpl.java
@@ -54,7 +54,7 @@ public class LogWtAttendanceRealizationDaoImpl extends IDAOImpl<LogWtAttendanceR
         query.append(" WHERE logAttendanceRealization.wtPeriodeId = :wtPeriodId ");
         
          return getCurrentSession().createQuery(query.toString())
-                    .setParameter("wtPeriodId", wtPeriodId.intValue())                   
+                    .setParameter("wtPeriodId", wtPeriodId)                   
                     .setMaxResults(maxResults).setFirstResult(firstResult)                
                     .setResultTransformer(Transformers.aliasToBean(TempAttendanceRealizationViewModel.class))
                     .list();
@@ -67,7 +67,7 @@ public class LogWtAttendanceRealizationDaoImpl extends IDAOImpl<LogWtAttendanceR
         query.append(" WHERE logAttendanceRealization.wtPeriodeId = :wtPeriodId ");
         
           Query hbm = getCurrentSession().createQuery(query.toString())
-                        .setParameter("wtPeriodId", wtPeriodId.intValue());
+                        .setParameter("wtPeriodId", wtPeriodId);
         return Long.valueOf(hbm.uniqueResult().toString());
     }
